added tailwind and added pages for redirection
