+++ src/components/pages/Signin.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Signin = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Signin
