+++ src/components/pages/Signup.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Signup = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Signup
